refactor(web): co-locate viavai types with components
